NewsPaper/requirements.txt
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
astroid==2.15.5
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
amqp==5.1.1
APScheduler==3.10.1
asgiref==3.6.0
astroid==2.15.5
billiard==4.1.0
celery==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.1
django-allauth==0.54.0
django-apscheduler==0.6.2
django-filter==23.2
django-modeltranslation==0.18.11
djangorestframework==3.14.0
idna==3.4
isort==5.12.0
kombu==5.3.1
lazy-object-proxy==1.9.0
Markdown==3.4.4
mccabe==0.7.0
oauthlib==3.2.2
platformdirs==3.8.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.7.0
pylint==2.17.4
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0

